backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
# faster-whisper is no longer needed as we use ffmpeg 8.0 built-in whisper
# faster-whisper
edge-tts
# moviepy often has issues, let's stick to direct ffmpeg-python or subprocess calls for stability if possible, but moviepy is easier for simple cutting.
moviepy
ffmpeg-python
openai
python-dotenv
requests
aiofiles
torch
torchvision
torchaudio
tenacity
pydantic-settings
pydub
imageio-ffmpeg
